Fix phone number overflow and simplify navigation
- Change Employee.phone_num from IntegerField to BigIntegerField to support 10-digit phone numbers
- Add migration 0005_alter_employee_phone_num.py for database schema update
- Simplify navigation bar by organizing items into logical dropdown menus (7 main items instead of 13+)
- Fix integer out of range error when creating sample employees

This resolves deployment errors and improves UX with cleaner navigation.

+++ emp_app/models.py
@@ -23,7 +23,7 @@ class Employee(models.Model):
     salary = models.IntegerField(default=0)
     bonus = models.IntegerField(default=0)
     role = models.ForeignKey(Role, on_delete=models.CASCADE)
-    phone_num = models.IntegerField(default=0)
+    phone_num = models.BigIntegerField(default=0)
     hire_date = models.DateField()
 
     def __str__(self):
